tweak build script for PRs

+++ _bin/build.sh
@@ -1,13 +1,9 @@
 #!/bin/bash
 
-# only proceed script when started not by pull request (PR)
-if [ $TRAVIS_PULL_REQUEST == "true" ]; then
-  echo "this is PR, exiting"
-  exit 0
-fi
-
 # enable error reporting to the console
 set -e
+
+echo $TRAVIS_PULL_REQUEST
 
 npm install bower
 bower install
@@ -17,19 +13,23 @@ echo $TRAVIS_BUILD_NUMBER > _includes/travis_build_number.txt
 # build site with jekyll, by default to `_site' folder
 bundle exec jekyll build
 
-# cleanup
-rm -rf gh-pages
+# Only deploy if not PR
+if $TRAVIS_PULL_REQUEST = false
+  then
+  # cleanup
+  rm -rf gh-pages
 
-git clone -b gh-pages https://${GH_TOKEN}@github.com/newtheatre/history-project.git gh-pages
+  git clone -b gh-pages https://${GH_TOKEN}@github.com/newtheatre/history-project.git gh-pages
 
-# copy generated HTML site to built branch
-cp -R _site/* gh-pages
+  # copy generated HTML site to built branch
+  cp -R _site/* gh-pages
 
-# commit and push generated content to built branch
-# since repository was cloned in write mode with token auth - we can push there
-cd gh-pages
-git config user.email "webmaster@newtheatre.org.uk"
-git config user.name "ntbot"
-git add -A .
-git commit -a -m "Travis Build $TRAVIS_BUILD_NUMBER"
-git push --quiet origin gh-pages > /dev/null 2>&1 # Hiding all the output from git push command, to prevent token leak.
+  # commit and push generated content to built branch
+  # since repository was cloned in write mode with token auth - we can push there
+  cd gh-pages
+  git config user.email "webmaster@newtheatre.org.uk"
+  git config user.name "ntbot"
+  git add -A .
+  git commit -a -m "Travis Build $TRAVIS_BUILD_NUMBER"
+  git push --quiet origin gh-pages > /dev/null 2>&1 # Hiding all the output from git push command, to prevent token leak.
+fi
